+++ legends/saveslot.py
@@ -2,16 +2,15 @@
 
 """
 
-from collections.abc import MutableMapping
 from datetime import datetime, timedelta, timezone
 # pylint: disable-next=no-name-in-module
-from legends.constants import GSCharacter, GSItem
+from legends.constants import GSCharacter
 from legends.constants import ITEMS
+from legends.gameobjects import Inventory
 from legends.roster import Roster
 
 __all__ = [
-    'ticksToTimedelta', 'ticksToDatetime', 'SaveSlot', 'STLTimeStamps',
-    'Inventory'
+    'ticksToTimedelta', 'ticksToDatetime', 'SaveSlot', 'STLTimeStamps'
 ]
 
 def ticksToTimedelta(ticks):
@@ -59,8 +58,8 @@ class SaveSlot():
             player's inventory.
         favorites (list of legends.gameobjects.Character): A list of
             characters the player has marked as 'favorite'.
-        inventory (Inventory): The inventory associated with the save
-            slot.
+        inventory (legends.gameobjects.Inventory): The inventory
+            associated with the save slot.
 
     """
 
@@ -167,96 +166,3 @@ class STLTimeStamps():
         self.playDuration = ticksToTimedelta(
             int(save['{} playDuration'.format(slot)])
         )
-
-class Inventory(MutableMapping):
-    """A collection of items in STL.
-
-    The `Inventory` class is a dictionary-like data structure, mapping
-    each item in `ITEMS` to the quantity of that item that exists in the
-    player's inventory. Keys cannot be deleted. Instead, deleting a key
-    simply changes its value to 0. Iterating over an `Inventory` object
-    will skip over items that are either irrelevant to the `legends`
-    package, or are implemented elsewhere. The skipped items are
-    determined by the `hiddenItemIDs` and `hiddenCategories` attributes.
-    To iterate over all keys, simply iterate over `ITEMS`.
-    The `__len__()` method also does not consider these skipped items.
-
-    """
-
-    hiddenCategories = ['Token', 'PlayerAvatar', 'Emote']
-    """`list of str`: A list of category names, as they appear in the
-    `category` attribute of an `legends.constants.Item` instance, that
-    are of limited use or implemented elsewhere in the `legends`
-    package.
-    """
-
-    hiddenItemIDs = [
-        'Credits', 'Dilithium', 'Tritanium', 'Player XP', 'PvP Stamina',
-        'Alliance Stamina', 'EventPoint', 'PvP Chest Points',
-        'Shards Advanced', 'Shards Elite', 'Shards Credit',
-        'Shards Biomimetic', 'Shards Protomatter', 'Shards_Worf',
-        'Shards_McCoy', 'Dungeon Currency', 'Dungeon Stamina'
-    ]
-    """`list of str`: A list of item IDs, as they appear in `GSItem`,
-    that are of limited use or implemented elsewhere in the `legends`
-    package.
-    """
-
-    def __init__(self, initDict=None):
-        """The constructor initializes the `Inventory` instance with one
-        key for each item in `ITEMS`, and all values 0. If the
-        `initData` argument is given, it is used to initialize the
-        values.
-
-        Args:
-            initData (dict): {`str`:`int`} A dictionary mapping item
-                IDs, as they appear in `GSItem`, to nonnegative
-                integers. Used to initialize the quantities in the
-                `Inventory` instance.
-
-        """
-        self._data = {}
-        for itemID in GSItem:
-            self._data[itemID] = 0
-        initDict = {} if initDict is None else initDict
-        for itemID, qty in initDict.items():
-            self._data[itemID] = qty
-
-    def __getitem__(self, item):
-        return self._data[item.itemID]
-
-    def __setitem__(self, item, qty):
-        self._data[item.itemID] = qty
-
-    def __delitem__(self, item):
-        self._data[item.itemID] = 0
-
-    def __iter__(self):
-        for itemID in self._data:
-            if not self._hidden(itemID):
-                yield ITEMS[itemID]
-
-    def __len__(self):
-        count = 0
-        for itemID in self._data:
-            if not self._hidden(itemID):
-                count += 1
-        return count
-
-    def _hidden(self, itemID):
-        if itemID in self.hiddenItemIDs:
-            return True
-        if ITEMS[itemID].category in self.hiddenCategories:
-            return True
-        return False
-
-    def keysByCategory(self, category):
-        """Returns an iterator over all keys that match the given
-        category, skipping any keys that are skipped during normal
-        iteration.
-
-        Args:
-            category (str): The category to iterate over.
-
-        """
-        return (item for item in self if item.category == category)